+++ vtr-ui/src/app/data-models/device/oled-power-settings.model.ts
@@ -1,0 +1,7 @@
+export class OledPowerSettings {
+	constructor(
+		public isCapable: boolean,
+		public taskBarDimmerValue: number,
+		public backgroundDimmerValue: number,
+		public displayDimmerValue: number) { }
+}