+++ vtr-ui/src/environments/environment.mock.ts
@@ -14,6 +14,5 @@ export const environment: IEnvironment = {
 	isServerSwitchEnabled: true,
 	spPaymentProcessApiRoot: 'https://uatpcsupport.lenovo.com/upgradewarranty?',
 	spGetOrdersApiRoot: 'https://uatpcsupport.lenovo.com/api/v4/upsell/smart/getorders?serialNumber=',
-	spPnListKey: 'U2FsdGVkX1+X2TVVOmuZWob7GGBAX0MqhbFG7Py5rjlN9MWpGuBkRkz23xBRr9xZllutntrWS2FDYJ9/CW0JIK8DaxmmShtyQEe+aqm9jYk='
-
+	spPnListKey: 'WyI1V1MwWDU4NjcxIiwiNVdTMFg1ODY3M'
 };